AI-Chatbot/requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
gtts==2.5.1
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
